refactor(client): client 패키지 생성 후 Player 클래스 이전
