+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import client.Player;
+
 public class Application {
     public static void main(String[] args) {
         Player p = new Player();
